updates to routes
