+++ src/components/Side/side.js
@@ -1,8 +1,15 @@
 import React from 'react';
-import Input from './Input/input.js';
+import Input from './input.js';
+import Search from './search.js';
+import Output from './output.js';
+
 
 export default function Side() {
-    const side = (
+  return(
+    <div class="side">
+      <Search/>
       <Input/>
-    );
+      <Output/>
+    </div>
+  );
 }
