+++ src/styles.css
@@ -9,6 +9,11 @@ h2, h3 {
     text-transform: uppercase;
 }
 
+
+ul {
+  display: flex;
+  flex-direction: row;
+}
 .page {
     margin: 0 auto;
     width: 100%;
@@ -20,7 +25,7 @@ h2, h3 {
 .header {
     background-color: #F6C555;
     width: auto;
-    height: 10vh;
+    height: 20%;
 }
 
 .side {
@@ -99,5 +104,4 @@ ul{
   font-size: 10px;
   text-align: left;
   list-style-type: none;
-
 }
